PcapCrystal/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "asyncio>=4.0.0",
    "diskcache>=5.6.3",
    "geoip2>=5.1.0",
    "httpx>=0.28.1",
    "networkx>=3.5",
    "numpy>=2.3.2",
    "pandas>=2.3.2",
    "plotly>=6.3.0",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
    "scapy>=2.6.1",
    "scikit-learn>=1.7.1",
    "streamlit>=1.48.1",
    "tenacity>=9.1.2",
]
